+++ sqlparse/__init__.py
@@ -9,14 +9,13 @@
 __version__ = '0.1.4'
 
 
-class SQLParseError(Exception):
-    """Base class for exceptions in this module."""
-
-
 # Setup namespace
 from sqlparse import engine
 from sqlparse import filters
 from sqlparse import formatter
+
+# Deprecated in 0.1.5. Will be removed in 0.2.0
+from sqlparse.exceptions import SQLParseError
 
 
 def parse(sql):