+++ sqlparse/exceptions.py
@@ -1,0 +1,11 @@
+# Copyright (C) 2012 Andi Albrecht, albrecht.andi@gmail.com
+#
+# This module is part of python-sqlparse and is released under
+# the BSD License: http://www.opensource.org/licenses/bsd-license.php.
+
+"""Exceptions used in this package."""
+
+class SQLParseError(Exception):
+    """Base class for exceptions in this module."""
+
+