+++ docs/source/changes.rst
@@ -3,5 +3,11 @@
 Changes in python-sqlparse
 ==========================
 
+Upcoming Deprecations
+---------------------
+
+* ``sqlparse.SQLParseError`` is deprecated (version 0.1.5), use
+  ``sqlparse.exceptions.SQLParseError`` instead.
+
 .. include:: ../../CHANGES